moved out fulltrim and languages
