+++ src/config/languages.js
@@ -1,0 +1,79 @@
+// Language/subdomain mappings for Wikipedia
+export const subdomains = {
+	"aa": "Afar", "ab": "Abkhaz", "ace": "Acehnese", "ady": "Adyghe", "af": "Afrikaans",
+	"ak": "Akan", "als": "Alemannic", "alt": "Southern Altai", "am": "Amharic", "ami": "Amis",
+	"an": "Aragonese", "ang": "Old English", "ann": "Obolo", "anp": "Angika", "ar": "Arabic",
+	"arc": "Aramaic", "ary": "Moroccan Arabic", "arz": "Egyptian Arabic", "as": "Assamese", "ast": "Asturian",
+	"atj": "Atikamekw", "av": "Avar", "avk": "Kotava", "awa": "Awadhi", "ay": "Aymara",
+	"az": "Azerbaijani", "azb": "South Azerbaijani", "ba": "Bashkir", "ban": "Balinese", "bar": "Bavarian",
+	"bat-smg": "Samogitian", "bbc": "Toba Batak", "bcl": "Central Bikol", "bdr": "Bajau Sama", "be": "Belarusian",
+	"be-tarask": "Belarusian (Classical)", "bew": "Betawi", "bg": "Bulgarian", "bh": "Bhojpuri", "bi": "Bislama",
+	"bjn": "Banjarese", "blk": "Pa'O", "bm": "Bambara", "bn": "Bengali", "bo": "Tibetan",
+	"bpy": "Bishnupriya Manipuri", "br": "Breton", "bs": "Bosnian", "btm": "Mandailing Batak", "bug": "Buginese",
+	"bxr": "Buryat", "ca": "Catalan", "cbk-zam": "Chavacano", "cdo": "Eastern Min", "ce": "Chechen",
+	"ceb": "Cebuano", "ch": "Chamorro", "cho": "Choctaw", "chr": "Cherokee", "chy": "Cheyenne",
+	"ckb": "Sorani Kurdish", "co": "Corsican", "cr": "Cree", "crh": "Crimean Tatar", "cs": "Czech",
+	"csb": "Kashubian", "cu": "Old Church Slavonic", "cv": "Chuvash", "cy": "Welsh", "da": "Danish",
+	"dag": "Dagbani", "de": "German", "dga": "Dagaare", "din": "Dinka", "diq": "Zazaki",
+	"dsb": "Lower Sorbian", "dtp": "Dusun", "dty": "Doteli", "dv": "Maldivian", "dz": "Dzongkha",
+	"ee": "Ewe", "el": "Greek", "eml": "Emilian–Romagnol", "en": "English", "eo": "Esperanto",
+	"es": "Spanish", "et": "Estonian", "eu": "Basque", "ext": "Extremaduran", "fa": "Persian",
+	"fat": "Fante", "ff": "Fula", "fi": "Finnish", "fiu-vro": "Võro", "fj": "Fijian",
+	"fo": "Faroese", "fon": "Fon", "fr": "French", "frp": "Franco-Provençal", "frr": "North Frisian",
+	"fur": "Friulian", "fy": "West Frisian", "ga": "Irish", "gag": "Gagauz", "gan": "Gan",
+	"gcr": "Guianan Creole", "gd": "Scottish Gaelic", "gl": "Galician", "glk": "Gilaki", "gn": "Guarani",
+	"gom": "Konkani", "gor": "Gorontalo", "got": "Gothic", "gpe": "Ghanaian Pidgin English", "gu": "Gujarati",
+	"guc": "Wayuu", "gur": "Gurenne", "guw": "Gun", "gv": "Manx", "ha": "Hausa", "hak": "Hakka",
+	"haw": "Hawaiian", "he": "Hebrew", "hi": "Hindi", "hif": "Fiji Hindi", "ho": "Hiri Motu",
+	"hr": "Croatian", "hsb": "Upper Sorbian", "ht": "Haitian Creole", "hu": "Hungarian", "hy": "Armenian",
+	"hyw": "Western Armenian", "hz": "Herero", "ia": "Interlingua", "iba": "Iban", "id": "Indonesian",
+	"ie": "Interlingue", "ig": "Igbo", "igl": "Igala", "ii": "Sichuan Yi", "ik": "Inupiaq",
+	"ilo": "Ilocano", "inh": "Ingush", "io": "Ido", "is": "Icelandic", "it": "Italian",
+	"iu": "Inuktitut", "ja": "Japanese", "jam": "Jamaican Patois", "jbo": "Lojban", "jv": "Javanese",
+	"ka": "Georgian", "kaa": "Karakalpak", "kab": "Kabyle", "kbd": "Kabardian", "kbp": "Kabiye",
+	"kcg": "Tyap", "kg": "Kongo", "kge": "Komering", "ki": "Kikuyu", "kj": "Kwanyama",
+	"kk": "Kazakh", "kl": "Greenlandic", "km": "Khmer", "kn": "Kannada", "knc": "Central Kanuri",
+	"ko": "Korean", "koi": "Komi-Permyak", "kr": "Kanuri", "krc": "Karachay-Balkar", "ks": "Kashmiri",
+	"ksh": "Ripuarian", "ku": "Kurdish", "kus": "Kusaal", "kv": "Komi", "kw": "Cornish",
+	"ky": "Kyrgyz", "la": "Latin", "lad": "Judaeo-Spanish", "lb": "Luxembourgish", "lbe": "Lak",
+	"lez": "Lezgian", "lfn": "Lingua Franca Nova", "lg": "Luganda", "li": "Limburgish", "lij": "Ligurian",
+	"lld": "Ladin", "lmo": "Lombard", "ln": "Lingala", "lo": "Lao", "lrc": "Northern Luri",
+	"lt": "Lithuanian", "ltg": "Latgalian", "lv": "Latvian", "mad": "Madurese", "mai": "Maithili",
+	"map-bms": "Banyumasan", "mdf": "Moksha", "mg": "Malagasy", "mh": "Marshallese", "mhr": "Meadow Mari",
+	"mi": "Maori", "min": "Minangkabau", "mk": "Macedonian", "ml": "Malayalam", "mn": "Mongolian",
+	"mni": "Meitei", "mnw": "Mon", "mo": "Moldovan", "mos": "Moore", "mr": "Marathi",
+	"mrj": "Hill Mari", "ms": "Malay", "mt": "Maltese", "mus": "Muscogee", "mwl": "Mirandese",
+	"my": "Burmese", "myv": "Erzya", "mzn": "Mazanderani", "na": "Nauruan", "nah": "Nahuatl",
+	"nap": "Neapolitan", "nds": "Low German", "nds-nl": "Dutch Low Saxon", "ne": "Nepali",
+	"new": "Newar", "ng": "Ndonga", "nia": "Nias", "nl": "Dutch", "nn": "Norwegian (Nynorsk)",
+	"no": "Norwegian", "nostalgia": "Nostalgia", "nov": "Novial", "nqo": "N'Ko", "nr": "Southern Ndebele",
+	"nrm": "Norman", "nso": "Northern Sotho", "nup": "Nupe", "nv": "Navajo", "ny": "Chewa",
+	"oc": "Occitan", "olo": "Livvi-Karelian", "om": "Oromo", "or": "Odia", "os": "Ossetian",
+	"pa": "Punjabi", "pag": "Pangasinan", "pam": "Kapampangan", "pap": "Papiamento", "pcd": "Picard",
+	"pcm": "Nigerian Pidgin", "pdc": "Pennsylvania Dutch", "pfl": "Palatine German", "pi": "Pali",
+	"pih": "Norfuk", "pl": "Polish", "pms": "Piedmontese", "pnb": "Western Punjabi", "pnt": "Pontic",
+	"ps": "Pashto", "pt": "Portuguese", "pwn": "Paiwan", "qu": "Quechua", "rki": "Rakhine",
+	"rm": "Romansh", "rmy": "Romani", "rn": "Kirundi", "ro": "Romanian", "roa-rup": "Aromanian",
+	"roa-tara": "Tarantino", "rsk": "Pannonian Rusyn", "ru": "Russian", "ru-sib": "Siberian",
+	"rue": "Rusyn", "rw": "Kinyarwanda", "sa": "Sanskrit", "sah": "Yakut", "sat": "Santali",
+	"sc": "Sardinian", "scn": "Sicilian", "sco": "Scots", "sd": "Sindhi", "se": "Northern Sami",
+	"sep11": "September 11", "sg": "Sango", "sh": "Serbo-Croatian", "shi": "Shilha", "shn": "Shan",
+	"si": "Sinhala", "simple": "Simple English", "sk": "Slovak", "skr": "Saraiki", "sl": "Slovene",
+	"sm": "Samoan", "smn": "Inari Sami", "sn": "Shona", "so": "Somali", "sq": "Albanian",
+	"sr": "Serbian", "srn": "Sranan Tongo", "ss": "Swazi", "st": "Sotho", "stq": "Saterland Frisian",
+	"su": "Sundanese", "sv": "Swedish", "sw": "Swahili", "syl": "Sylheti", "szl": "Silesian",
+	"szy": "Sakizaya", "ta": "Tamil", "tay": "Atayal", "tcy": "Tulu", "tdd": "Tai Nuea",
+	"te": "Telugu", "ten": "10", "test": "Test", "test2": "Test2", "tet": "Tetum",
+	"tg": "Tajik", "th": "Thai", "ti": "Tigrinya", "tig": "Tigre", "tk": "Turkmen",
+	"tl": "Tagalog", "tlh": "Klingon", "tly": "Talysh", "tn": "Tswana", "to": "Tongan",
+	"tokipona": "Toki Pona", "tpi": "Tok Pisin", "tr": "Turkish", "trv": "Seediq",
+	"ts": "Tsonga", "tt": "Tatar", "tum": "Tumbuka", "tw": "Twi", "ty": "Tahitian",
+	"tyv": "Tuvan", "udm": "Udmurt", "ug": "Uyghur", "uk": "Ukrainian", "ur": "Urdu",
+	"uz": "Uzbek", "ve": "Venda", "vec": "Venetian", "vep": "Veps", "vi": "Vietnamese",
+	"vls": "West Flemish", "vo": "Volapük", "wa": "Walloon", "war": "Waray", "wo": "Wolof",
+	"wuu": "Wu", "xal": "Kalmyk", "xh": "Xhosa", "xmf": "Mingrelian", "yi": "Yiddish",
+	"yo": "Yoruba", "za": "Zhuang", "zea": "Zeelandic", "zgh": "Moroccan Amazigh", "zh": "Chinese",
+	"zh-classical": "Classical Chinese", "zh-min-nan": "Southern Min", "zh-yue": "Cantonese",
+	"zu": "Zulu"
+};
+
